Move transformers registry to definitions.js
